complete-mockito2-course-for-beginners/src/main/java/com/vsushko/GenericService.java
package com.vsushko;

/**
 * 1) Constructor based injection
 * 2) Setter based injection
 * 3) Field based injection
 *
 */
public class GenericService {

    private LibraryService.DAO dao;
    private LoggerService outLogger;
    private LoggerService errLogger;


    public Book fetchBookById(Integer id) {
        System.out.println("GenericService.foo");

        final Book book1 = dao.fetchBookById(id);
        final Book book = dao.fetchBookByTitle("foo");
        outLogger.logToOut(book.getTitle());
        errLogger.logToErr(book.getTitle());
        return book;
    }

    public Book fetchBook(Integer id, String title) {
        System.out.println("GenericService.fetchBook");

        Book book = dao.fetchBookById(id);
        if (book==null) {
            book = dao.fetchBookByTitle(title);
        }
        outLogger.logToOut(book.getTitle());
        errLogger.logToErr(book.getTitle());
        return book;
    }
}
